Refactor slice changes
